rl/sb3_app/pyproject.toml
[tool.poetry]
name = "sb3-app"
version = "0.1.0"
description = "SB3 runner for the board game RL project"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.11"
stable-baselines3 = {extras = ["extra"], version = "^2.6.0"}
board-game-env-lib = {path = "../board_game_lib", develop = true}
pyyaml = "^6.0.2"
pydantic = "^2.11.7"
torch = {version = ">=2.3.0", source = "pytorch_cuda"}
gradio = "^5.34.2"
ruamel-yaml = "^0.18.14"
tensorflow = "2.15.0"  # ✅ Add this line


[[tool.poetry.source]]
name = "pytorch_cuda"
url = "https://download.pytorch.org/whl/cu121"
priority = "explicit"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
